backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
bcrypt==4.0.1
marshmallow==3.19.0
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0